MA-4: refactoring
