+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MovieApp">
-        <activity android:name=".ui.ActivityMain">
+        <activity android:name=".ui.scenes.ActivityMain">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
