--- app/src/main/java/ru/movie/app/ui/views/Context.kt
@@ -1,6 +1,0 @@
-package ru.movie.app.ui.views
-
-import android.content.Context
-import android.util.TypedValue
-
-fun Context.dpToPx(dp: Int): Float = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp.toFloat(), this.resources.displayMetrics)
