+++ app/src/main/res/drawable/shape_rounded_with_shadow.xml
@@ -1,11 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="@color/background_shadow" />
-            <corners android:radius="6dp" />
-        </shape>
-    </item>
+    <item android:drawable="@drawable/shape_shadow" android:top="23dp"/>
     <item android:bottom="3dp">
         <shape android:shape="rectangle">
             <solid android:color="@color/background" />
