+++ app/src/main/res/drawable/shape_shadow.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <solid android:color="@color/background_shadow"/>
+    <corners android:bottomRightRadius="6dp" android:bottomLeftRadius="6dp"/>
+</shape>
