+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".ui.ActivityMain">
+    tools:context=".ui.scenes.ActivityMain">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -43,7 +43,7 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="16dp"
             android:layout_marginBottom="12dp"
-            android:background="@drawable/shape_rounded"
+            android:background="@drawable/shape_rounded_with_shadow"
             android:paddingHorizontal="2dp"
             android:paddingTop="4dp"
             android:paddingBottom="7dp"
